// packages/core/src/generated/ClipSource.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Generator } from "./Generator";
import type { MediaId } from "./MediaId";
import type { Timeline } from "./Timeline";

export type ClipSource = { "kind": "media", media: MediaId, } | { "kind": "generator", generator: Generator, } | { "kind": "compound", timeline: Timeline, };