// packages/core/src/generated/Generator.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { JsonValue } from "./serde_json/JsonValue";

export type Generator = { "type": "text", content: string, style: { [key in string]: JsonValue }, } | { "type": "solid", color: string, } | { "type": "shape", shape: string, color: string, } | { "type": "gradient", from: string, to: string, angle: number, } | { "type": "countdown", from_seconds: number, };