// packages/core/src/generated/Keyframe.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Interp } from "./Interp";
import type { ParamValue } from "./ParamValue";
import type { Time } from "./Time";

export type Keyframe = { time: Time, value: ParamValue, interp: Interp, handleIn?: [number, number] | null, handleOut?: [number, number] | null, };